controlplane: Refactor servers
This commit decouples controlplane server handlers from a specific server.
This will allow customization of handler registration.
For example, registering with or without REST API handlers.
