--- pkg/controlplane/server/http/heartbeat.go
@@ -1,31 +1,0 @@
-// Copyright 2023 The ClusterLink Authors.
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package http
-
-import (
-	"net/http"
-
-	"github.com/clusterlink-net/clusterlink/pkg/controlplane/api"
-)
-
-func (s *Server) addHeartbeatHandler() {
-	r := s.Router()
-
-	r.Get(api.HeartbeatPath, s.Heartbeat)
-}
-
-// Heartbeat returns a response for heartbeat checks from remote peers.
-func (s *Server) Heartbeat(w http.ResponseWriter, _ *http.Request) {
-	w.WriteHeader(http.StatusOK)
-}
