+++ pkg/util/grpc/server.go
@@ -34,7 +34,7 @@ func (s *Server) GetGRPCServer() *grpc.Server {
 	return s.server
 }
 
-// Start the server.
+// Serve starts the server.
 func (s *Server) Start() error {
 	return s.server.Serve(s.GetListener())
 }
@@ -52,8 +52,8 @@ func (s *Server) GracefulStop() error {
 }
 
 // NewServer returns a new server.
-func NewServer(name string, tlsConfig *tls.Config) Server {
-	return Server{
+func NewServer(name string, tlsConfig *tls.Config) *Server {
+	return &Server{
 		Listener: tcp.NewListener(name),
 		server:   grpc.NewServer(grpc.Creds(credentials.NewTLS(tlsConfig))),
 	}
